feat: add eager loading for user relations in rating and friend request entities
The changes in this commit add eager loading for user relations in the rating and friend request entities. This means that when retrieving a rating or a friend request, the associated user object will also be loaded. This improves performance by reducing the number of database queries needed to fetch the related user data.