GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Remove unused dependencies (refs/gitbutler/remove-unused-dependencies)
   - .github/workflows/release.yml
   - package.json
   - release.config.mjs
   - yarn.lock

Your previous branch was: refs/heads/developer

The sha for that commit was: bac9153a9007fb4801b3f7106095155858b72a93

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch